### 2022-07-25 #4604 @editorialbot
[PRE REVIEW]: PACMAN: A pipeline to reduce and analyze Hubble Wide Field Camera 3 IF Grism data
**Submitting author:** @sebastian-zieba (<a href="http://orcid.org/0000-0003-0562-6750">Sebastian Zieba</a>)
**Repository:** https://github.com/sebastian-zieba/PACMAN
**Branch with paper.md** (empty if default branch):
**Version:** v0.1-alpha
