### 2022-08-09 #4604 @xuanxu
:wave: @nespinoza, @noraeisner would you be willing to review this paper? I think your expertise with Exoplanets and Python makes both of you a great fit for this submission.
