### 2022-08-17 #4604 @aureliocarnero
Dear @xuanxu, unfortunately I cannot either, sorry. Hopefully next call I
can

El sáb., 13 ago. 2022 17:05, Hannu Parviainen ***@***.***>
escribió:

> Sorry @xuanxu <https://github.com/xuanxu>, I also have my hands full at
> the moment :/ I don't think I'd have the time before September.
>
