### 2022-09-01 7efad53
fixed necessity to manually change cjs in formatter
