+++ src/pacman/lib/formatter.py
@@ -59,36 +59,13 @@ class FormatParams:
             self.log_rho = params[data.par_order['log_rho'] * data.nvisit:(1 + data.par_order['log_rho']) * data.nvisit]
             self.log_sigma = params[data.par_order['log_sigma'] * data.nvisit:(1 + data.par_order['log_sigma']) * data.nvisit]
         elif 'constants_cj' in data.s30_myfuncs:
-            self.c0 = params[data.par_order['c0']*data.nvisit:(1 + data.par_order['c0'])*data.nvisit]
-            self.c1 = params[data.par_order['c1']*data.nvisit:(1 + data.par_order['c1'])*data.nvisit]
-            self.c2 = params[data.par_order['c2']*data.nvisit:(1 + data.par_order['c2'])*data.nvisit]
-            self.c3 = params[data.par_order['c3']*data.nvisit:(1 + data.par_order['c3'])*data.nvisit]
-            self.c4 = params[data.par_order['c4']*data.nvisit:(1 + data.par_order['c4'])*data.nvisit]
-            self.c5 = params[data.par_order['c5']*data.nvisit:(1 + data.par_order['c5'])*data.nvisit]
-            self.c6 = params[data.par_order['c6']*data.nvisit:(1 + data.par_order['c6'])*data.nvisit]
-            self.c7 = params[data.par_order['c7']*data.nvisit:(1 + data.par_order['c7'])*data.nvisit]
-            self.c8 = params[data.par_order['c8']*data.nvisit:(1 + data.par_order['c8'])*data.nvisit]
-            self.c9 = params[data.par_order['c9']*data.nvisit:(1 + data.par_order['c9'])*data.nvisit]
-            self.c10 = params[data.par_order['c10']*data.nvisit:(1 + data.par_order['c10'])*data.nvisit]
-            self.c11 = params[data.par_order['c11']*data.nvisit:(1 + data.par_order['c11'])*data.nvisit]
-            self.c12 = params[data.par_order['c12']*data.nvisit:(1 + data.par_order['c12'])*data.nvisit]
-            self.c13 = params[data.par_order['c13']*data.nvisit:(1 + data.par_order['c13'])*data.nvisit]
-            self.c14 = params[data.par_order['c14']*data.nvisit:(1 + data.par_order['c14'])*data.nvisit]
-            self.c15 = params[data.par_order['c15']*data.nvisit:(1 + data.par_order['c15'])*data.nvisit]
-            self.c16 = params[data.par_order['c16']*data.nvisit:(1 + data.par_order['c16'])*data.nvisit]
-            self.c17 = params[data.par_order['c17']*data.nvisit:(1 + data.par_order['c17'])*data.nvisit]
-            self.c18 = params[data.par_order['c18']*data.nvisit:(1 + data.par_order['c18'])*data.nvisit]
-            self.c19 = params[data.par_order['c19']*data.nvisit:(1 + data.par_order['c19'])*data.nvisit]
-            self.c20 = params[data.par_order['c20']*data.nvisit:(1 + data.par_order['c20'])*data.nvisit]
-            self.c21 = params[data.par_order['c21']*data.nvisit:(1 + data.par_order['c21'])*data.nvisit]
-            self.c22 = params[data.par_order['c22']*data.nvisit:(1 + data.par_order['c22'])*data.nvisit]
-            self.c23 = params[data.par_order['c23']*data.nvisit:(1 + data.par_order['c23'])*data.nvisit]
-            self.c24 = params[data.par_order['c24']*data.nvisit:(1 + data.par_order['c24'])*data.nvisit]
-            #self.c25 = params[data.par_order['c25']*data.nvisit:(1 + data.par_order['c25'])*data.nvisit]
[…]
