### 2022-09-01 db6b7ef
added uncertainty multiplier as a fit function
