### 2022-09-01 98c47de
cleaned up least_squares and now loads white_sys from path in pcf
