+++ src/pacman/lib/least_squares.py
@@ -5,25 +5,31 @@ from . import plots
 from .formatter import PrintParams
 import pickle
 from astropy.stats import sigma_clip
+from . import util
 
 
-def residuals(params, data, model, fjac=None):			
+def residuals(params, data, model, fjac=None):
     fit = model.fit(data, params)
     return [0, fit.resid/data.err]
 
 
 def lsq_fit(fit_par, data, meta, model, myfuncs, noclip=False):
-    #TODO: noclip = True should be standard
+    """
+    Runs the least square routine using MPFIT
+    """
+    util.create_res_dir(meta) # creates the lsq directory
+
+    # TODO: noclip = True should be standard
     nvisit = data.nvisit 
     npar = len(data.parnames)*nvisit
 
-    #initializes least squares fit parameters
+    # initializes least squares fit parameters
     parinfo = [{'value':0, 'fixed':0, 'limited':[0,0,], 'limits':[0.0,0.0], 
                 'step':0.0} for j in range(npar)]
     params_s = []
 
-    #loops through parameters and visits
-    #sets initial guess, step size, tie, bounds
+    # loops through parameters and visits
+    # sets initial guess, step size, tie, bounds
 
     ii=0
     for i in range(int(len(data.parnames))):
@@ -58,19 +64,18 @@ def lsq_fit(fit_par, data, meta, model, myfuncs, noclip=False):
                 params_s.append(fit_par['value'][ii])
                 ii = ii + 1
 
-
     params_s = np.array(params_s)
     if meta.save_raw_lc_plot: plots.plot_raw(data, meta)
     fa = {'data':data, 'model':model}
 
     if ('divide_white' in data.s30_myfuncs) and meta.s30_fit_spec:
-        sys_vector = np.genfromtxt(meta.workdir + "/white_systematics.txt")
+        sys_vector = np.genfromtxt(meta.white_sys_path)
         data.all_sys = sys_vector
         #print("subtracting 2 from dof for divide-white")
         #data.nfree_param -= 2
         #data.dof += 2
 
-    print('\nRuns MPFIT... ')
+    print('Runs MPFIT... ')
     m = mpfit.mpfit(residuals, params_s, functkw=fa, parinfo = parinfo, quiet=True)
 
     if noclip == False:
@@ -84,43 +89,27 @@ def lsq_fit(fit_par, data, meta, model, myfuncs, noclip=False):
             print('Outlier Identified: ', len(clip_idx))
             print('Outlier idx: ', clip_idx)
             if meta.save_fit_lc_plot: plots.plot_fit_lc(data, model, meta)
-
-
-    #print(m.params)
-
-    #rescale error bars based on chi2
-    #print("rescaling error bars to get chi2red = 1")
-    #print("scale factor = ", np.sqrt(model.chi2red))
-    #print data.wavelength, np.sqrt(model.chi2red)
-    #data.err = data.err*np.sqrt(model.chi2red)
-    #m = mpfit.mpfit(residuals, params_s, functkw=fa, parinfo = parinfo, quiet=True)
-    #model = Model(m.params, data, flags)
     
     if m.errmsg: print("MPFIT error message", m.errmsg)
 
-
-    # if meta.run_output:
-    #     f = open(meta.run_out_name, "a")
-    #     print("{0:0.3f}".format(data.wavelength), \
-    #               "{0:0.6f}".format(m.params[data.par_order['rp']*nvisit]), \
-    #               "{0:0.6f}".format(m.perror[data.par_order['rp']*nvisit]),\
[…]
