### 2022-09-01 72d59dd
cleaned up mcmc and nested and made it compatible with uncmulti
