+++ src/pacman/lib/model.py
@@ -1,14 +1,13 @@
 import sys
-sys.path.insert(0, './models')
 import numpy as np
-from ..lib.formatter import FormatParams
 from ..lib.functions import Functions
+sys.path.insert(0, './models')
 
 
 def calc_astro(t, params, data, funcs, visit):
     flux = np.ones_like(t)
     for i, f in enumerate(funcs.astro): 
-        #selects parameters to pass to function
+        # selects parameters to pass to function
         funcparams = [params[j:j + data.nvisit] for j in funcs.astro_porder[i]]
         flux *= f(t, data, funcparams, visit)
     return flux 
@@ -16,8 +15,8 @@ def calc_astro(t, params, data, funcs, visit):
 
 def calc_sys(t, params, data, funcs, visit):
     flux = np.ones_like(t)
-    for i, f in enumerate(funcs.sys): 
-        #selects parameters to pass to function
+    for i, f in enumerate(funcs.sys):
+        # selects parameters to pass to function
         funcparams = [params[j: j + data.nvisit] for j in funcs.sys_porder[i]]
         flux *= f(t, data, funcparams, visit)
     return flux 
@@ -26,12 +25,11 @@ def calc_sys(t, params, data, funcs, visit):
 def calc_gp(idx, params, data, resid, funcs, visit):
     flux = np.ones(int(sum(idx)))
     for i, f in enumerate(funcs.gp):
-        #selects parameters to pass to function
+        # selects parameters to pass to function
         funcparams = [params[j + visit] for j in funcs.gp_porder[i]]
         gp, lnlike = f(idx, data, resid, funcparams)
         flux *= gp
     return [flux, lnlike]
-
 
 
 class Model:
@@ -53,20 +51,22 @@ class Model:
         self.chi2red = 0.
         self.rms = 0.
         self.rms_predicted = 1.0e6*np.sqrt(np.mean((data.err/data.flux)**2))
-        #self.rms_predicted = 1.0e6*np.sqrt(np.mean(np.sqrt((1./data.flux))**2)) #wrong because we binned over several pixels
+        print(f'The predicted rms is {self.rms_predicted:.2f}')
+        # self.rms_predicted = 1.0e6*np.sqrt(np.mean(np.sqrt((1./data.flux))**2))
+        # wrong because we binned over several pixels
         self.ln_like = 0.
         self.bic = 0.
+        self.bic_alt = 0.
         self.params = []
         self.myfuncs = Functions(data, myfuncs)
 
-
     def fit(self, data, params):
-        #loop over each observation
+        # loop over each observation
         for visit in range(data.nvisit):
-            #FIXME don't do this every time fit is run
+            # FIXME don't do this every time fit is run
             ind = data.vis_num == visit
 
-            #t = data.time[ind]
+            # t = data.time[ind]
             t = data.time
             per  = params[data.par_order['per']*data.nvisit + visit]
             t0  = params[data.par_order['t0']*data.nvisit + visit] + data.toffset
@@ -83,14 +83,16 @@ class Model:
         self.all_sys = data.flux/self.model_astro
         self.resid = data.flux - self.model
         self.norm_resid = self.resid/data.flux
-        self.chi2 = np.sum((self.resid/data.err)**2)		
+        self.chi2 = np.sum((self.resid/data.err)**2)
         self.chi2red = self.chi2/data.dof
         self.rms = 1.0e6*np.sqrt(np.mean((self.resid/data.flux)**2))
         if ('gp_sho' not in data.s30_myfuncs) and ('gp_matern32' not in data.s30_myfuncs):
             self.ln_like = (-0.5*(np.sum((self.resid/data.err)**2
[…]
