### 2022-09-01 cdd189e
cleaned up read_data
