+++ src/pacman/lib/read_data.py
@@ -15,16 +15,19 @@ class Data:
     """
     def __init__(self, data_file, meta, fit_par, clip_idx=[]):
         def format_prior_for_mcmc(self, meta, fit_par):
-                nvisit = int(meta.nvisit)
-                prior = []
+            """
+            Reads in the priors from the fit_par file.
+            """
+            nvisit = int(meta.nvisit)
+            prior = []
 
-                for i in range(len(fit_par)):
-                    if fit_par['fixed'][i].lower() == "false":
-                        prior.append([fit_par['prior'][i], float(fit_par['p1'][i]),
-                            float(fit_par['p2'][i])])
+            for i in range(len(fit_par)):
+                # TODO: IMPORTANT! Not only check if its a free parameters but also if it's used!!
+                if fit_par['fixed'][i].lower() == "false":
+                    prior.append([fit_par['prior'][i], float(fit_par['p1'][i]),
+                        float(fit_par['p2'][i])])
 
-                #print('prior, read_data', prior)
-                return prior
+            return prior
 
         #read in data
         d = ascii.read(data_file)
@@ -36,7 +39,7 @@ class Data:
         # Removes first exposure from each orbit
         # The index where the orbit changes is saved in meta.new_orbit_idx_sp
         if meta.remove_first_exp:
-            #removes first exposure from each orbit
+            # Removes first exposure from each orbit
             leave_ind = np.ones(len(d), bool)
             leave_ind[meta.new_orbit_idx_sp] = False
             d = d[leave_ind]
@@ -44,7 +47,7 @@ class Data:
             iorbit_sp = iorbit_sp[leave_ind]
             iexp_orb_sp = iexp_orb_sp[leave_ind]
             iexp_orb_sp -= 1 #because we removed iexp_orb = 0 and want the first one in an orbit to be 0 again
-            print('Removed {0} exposures because they were the first exposures in the orbit.'.format(sum(~leave_ind)))
+            print(f'Removed {sum(~leave_ind)} exposures because they were the first exposures in the orbit.')
         else:
             print('Leaving the first exposures in every orbit.')
 
@@ -53,25 +56,25 @@ class Data:
         # The NOT cumulative orbit list is saved in meta.iorbit_sp
         if len(meta.remove_which_orb) == 1:
             if meta.remove_first_orb:
-                #removes first exposure from each orbit
+                # Removes first exposure from each orbit
                 leave_ind = iorbit_sp != 0
                 d = d[leave_ind]
                 # Remove first orbit from meta.iorbit_sp too
                 iorbit_sp = iorbit_sp[leave_ind]
                 iexp_orb_sp = iexp_orb_sp[leave_ind]
-                print('Removed {0} exposures because they were the first orbit in the visit.'.format(sum(~leave_ind)))
+                print(f'Removed {sum(~leave_ind)} exposures because they were the first orbit in the visit.')
         elif len(meta.remove_which_orb) != 1 and meta.remove_first_orb:
             if meta.remove_first_orb:
                 masks_orb = []
                 for i in range(len(meta.remove_which_orb)):
                     masks_orb.append(iorbit_sp != meta.remove_which_orb[i])
-                #removes chosen orbits from each orbit
+                # Removes chosen orbits from each orbit
                 leave_ind = np.bitwise_and(*masks_orb)
                 d = d[leave_ind]
                 # Remove first orbit from meta.iorbit_sp too
                 iorbit_sp = iorbit_sp[leave_ind]
                 iexp_orb_sp = iexp_orb_sp[leave_ind]
[…]
