### 2022-09-01 d4d3b9f
setting up lists for final table
