+++ src/pacman/lib/util.py
@@ -14,10 +14,10 @@ from .sort_nicely import sort_nicely as sn
 import glob
 import pickle
 from scipy.signal import find_peaks
+from astropy.table import Table
 
 
 #s00
-
 def readfiles(meta):
     """
     Reads in the files saved in datadir and saves them into a list
@@ -599,6 +599,8 @@ def read_fitfiles(meta):
 
     print('Identified file(s) for fitting:', files)
 
+    meta.nfits = len(files)
+
     return files, meta
 
 
@@ -710,53 +712,39 @@ def make_lsq_rprs_txt(vals, errs, idxs, meta):
     """
     Saves the rprs vs wvl as a txt file as resulting from the lsq.
     """
-    if not os.path.isdir(meta.workdir + meta.fitdir + '/lsq_res'):
-        os.makedirs(meta.workdir + meta.fitdir + '/lsq_res')
     f_lsq = open(meta.workdir + meta.fitdir + '/lsq_res' + "/lsq_rprs.txt", 'w')
     rp_idx = np.where(np.array(meta.labels) == 'rp')[0][0]
     rprs_vals_lsq = [vals[ii][idxs[0][rp_idx]] for ii in range(len(vals))]
     rprs_errs_lsq = [errs[ii][idxs[0][rp_idx]] for ii in range(len(errs))]
-    file_header = ['wavelength (micron)', 'rprs', 'rprs_err', 'chi2red']
-    print("#{: <24} {: <25} {: <25} {: <25}".format(*file_header), file=f_lsq)
-    for row in zip(meta.wavelength_list, rprs_vals_lsq, rprs_errs_lsq, meta.chi2red_list):
-        print("{: <25} {: <25} {: <25} {: <25}".format(*row), file=f_lsq)
+    file_header = ['wavelength (micron)', 'rprs', 'rprs_err']
+    print("#{: <24} {: <25} {: <25}".format(*file_header), file=f_lsq)
+    for row in zip(meta.wavelength_list, rprs_vals_lsq, rprs_errs_lsq):
+        print("{: <25} {: <25} {: <25}".format(*row), file=f_lsq)
     f_lsq.close()
 
 
-def make_mcmc_rprs_txt(vals_mcmc, errs_lower_mcmc, errs_upper_mcmc, meta):
+def make_rprs_txt(vals, errs_lower, errs_upper, meta, fitter=None):
     """
-    Saves the rprs vs wvl as a txt file as resulting from the MCMC.
-    """
-    rp_idx = np.where(np.array(meta.labels) == 'rp')[0][0]
-    medians = np.array(vals_mcmc).T[rp_idx]
-    errors_lower = np.array(errs_lower_mcmc).T[rp_idx]
-    errors_upper = np.array(errs_upper_mcmc).T[rp_idx]
-    if not os.path.isdir(meta.workdir + meta.fitdir + '/mcmc_res'):
-        os.makedirs(meta.workdir + meta.fitdir + '/mcmc_res')
-    f_mcmc = open(meta.workdir + meta.fitdir + '/mcmc_res' + "/mcmc_rprs.txt", 'w')
-    file_header = ['wavelength (micron)', 'rprs', 'rprs_err_lower', 'rprs_err_upper', 'rms_pred', 'rms_obs']
-    print("#{: <24} {: <25} {: <25} {: <25}".format(*file_header), file=f_mcmc)
-    for row in zip(meta.wavelength_list, medians, errors_lower, errors_upper, meta.rms_pred_list, meta.rms_list_emcee):
-        print("{: <25} {: <25} {: <25} {: <25}".format(*row), file=f_mcmc)
-    f_mcmc.close()
-
-
[…]
