### 2022-09-01 55e349a
cleaned up s30
