+++ src/pacman/s30_run.py
@@ -31,128 +31,132 @@ def run30(eventlabel, workdir, meta=None):
 
     # Create directories for Stage 3 processing
     datetime = time.strftime('%Y-%m-%d_%H-%M-%S')
-    meta.fitdir = '/fit_' + datetime + '_' + meta.eventlabel
-    if not os.path.exists(meta.workdir + meta.fitdir):
-        os.makedirs(meta.workdir + meta.fitdir)
 
-    # Make fit_par nice...
+    # Create a directory for the white or the spectroscopic fit
+    if  meta.s30_fit_white:
+        meta.fitdir = f'/fit_white/fit_{datetime}_{meta.eventlabel}'
+    elif meta.s30_fit_spec:
+        meta.fitdir = f'/fit_spec/fit_{datetime}_{meta.eventlabel}'
+
+    if not os.path.exists(meta.workdir + meta.fitdir):
+        os.makedirs(meta.workdir + meta.fitdir, exist_ok=True)
+
+    # Make fit_par nicer by lining up the columns
     nice_fit_par.nice_fit_par(meta.workdir + "/fit_par.txt")
 
-    # Copy pcf and fit_par files
+    # Copy pcf and fit_par files into the new fitdirectory
     shutil.copy(meta.workdir + "/obs_par.pcf", meta.workdir + meta.fitdir)
     shutil.copy(meta.workdir + "/fit_par.txt", meta.workdir + meta.fitdir)
 
-    # reads in fit parameters
+    # Reads in fit parameters from the fit_par file
     #TODO: Check that fit_par is configured correctly. Eg initial value has to be within boundaries!
     fit_par = ascii.read(meta.workdir + "/fit_par.txt", Reader=ascii.CommentedHeader)
 
-    #read in the user wanted fit functions
+    # Read in the user wanted fit functions
     myfuncs = meta.s30_myfuncs
 
-    #store files to fit
+    # Indentify path of light curve files
     files, meta = util.read_fitfiles(meta)
 
+    # Prepare some lists for diagnostics
+    # TODO: Maybe make dict out of this?
     if meta.run_verbose:
         vals = []
         errs = []
         idxs = []
-
     if meta.run_mcmc:
         vals_mcmc       = []
         errs_lower_mcmc = []
         errs_upper_mcmc = []
-
     if meta.run_nested:
         vals_nested = []
         errs_lower_nested = []
         errs_upper_nested = []
 
-    meta.chi2red_list = []
-    meta.rms_pred_list = []
-    meta.rms_list = []
-    meta.rms_list_emcee = []
-    meta.rms_list_nested = []
+    meta = util.log_run_setup(meta)
 
+    # Loop through light curves
     for counter, f in enumerate(files):
         print('\n****** File: {0}/{1}'.format(counter+1, len(files)))
         meta.s30_file_counter = counter
-        time.sleep(1.1) #sleep to prevent overwriting of data if saved in the same second
+        time.sleep(1.01) #sleep to prevent overwriting of data if saved in the same second
         meta.run_file = f
         meta.fittime = time.strftime('%Y-%m-%d_%H-%M-%S')
 
         ### RUN LEAST SQUARES ###
         ## IF NO CLIPPING IS WISHED
-        if meta.run_clipiters == 0:
-            print('\n')
-            data = Data(f, meta, fit_par)
-            model = Model(data, myfuncs)
-            print('*STARTS LEAST SQUARED*')
-            data, model, params, m = lsq_fit(fit_par, data, meta, model, myfuncs, noclip=True) #not clipping
-            meta.chi2red_list.append(model.chi2red)
-            meta.rms_list.append(model.rms)
[…]
