+++ docs/source/api.rst
@@ -4,60 +4,60 @@ The Code (API)
 ===============
 
 
-reduction
-"""""""""""""""""""""""""""
+pacman
+""""""
 
-reduction.s00_table
-----------------------------------------
+s00_table
+---------
 .. automodule:: pacman.s00_table
     :members:
     :undoc-members:
     :show-inheritance:
 
-reduction.s01_horizons
-----------------------------------------
+s01_horizons
+------------
 .. automodule:: pacman.s01_horizons
     :members:
     :undoc-members:
     :show-inheritance:
 
-reduction.s02_barycorr
-----------------------------------------
+s02_barycorr
+------------
 .. automodule:: pacman.s02_barycorr
     :members:
     :undoc-members:
     :show-inheritance:
 
-reduction.s03_refspectra
-----------------------------------------
+s03_refspectra
+--------------
 .. automodule:: pacman.s03_refspectra
     :members:
     :undoc-members:
     :show-inheritance:
 
-reduction.s10_direct_images
--------------------------------------------
+s10_direct_images
+-----------------
 .. automodule:: pacman.s10_direct_images
     :members:
     :undoc-members:
     :show-inheritance:
 
-reduction.s20_extract
--------------------------------------------
+s20_extract
+-----------
 .. automodule:: pacman.s20_extract
     :members:
     :undoc-members:
     :show-inheritance:
 
-reduction.s21_bin_spectroscopic_lc
--------------------------------------------
+s21_bin_spectroscopic_lc
+------------------------
 .. automodule:: pacman.s21_bin_spectroscopic_lc
     :members:
     :undoc-members:
     :show-inheritance:
 
-reduction.s30_run
--------------------------------------------
+s30_run
+-------
 .. automodule:: pacman.s30_run
     :members:
     :undoc-members:
@@ -65,24 +65,24 @@ reduction.s30_run
 
 
 lib
-"""""""""""""""""""""""""""
+"""
 
 lib.util
-----------------------------------------
+--------
 .. automodule:: pacman.lib.util
     :members:
     :undoc-members:
     :show-inheritance:
 
 lib.read_pcf
-----------------------------------------
+------------
 .. automodule:: pacman.lib.read_pcf
     :members:
     :undoc-members:
     :show-inheritance:
 
 lib.manageevent
-----------------------------------------
+---------------
 .. automodule:: pacman.lib.manageevent
     :members:
     :undoc-members:
@@ -91,21 +91,21 @@ lib.manageevent
 
 
 lib.update_meta
-----------------------------------------
+---------------
 .. automodule:: pacman.lib.update_meta
     :members:
     :undoc-members:
     :show-inheritance:
 
 lib.suntimecorr
-----------------------------------------
+---------------
 .. automodule:: pacman.lib.suntimecorr
     :members:
     :undoc-members:
     :show-inheritance:
 
 lib.plots
-----------------------------------------
+---------
 .. automodule:: pacman.lib.plots
     :members:
     :undoc-members:
@@ -114,7 +114,7 @@ lib.plots
 
 
 lib.stellar_spectrum
-----------------------------------------
+--------------------
 .. automodule:: pacman.lib.stellar_spectrum
     :members:
     :undoc-members:
@@ -124,14 +124,14 @@ lib.stellar_spectrum
 
 
 lib.gaussfitter
-----------------------------------------
+---------------
 .. automodule:: pacman.lib.gaussfitter
     :members:
     :undoc-members:
     :show-inheritance:
 
 lib.mpfit
-----------------------------------------
+---------
 .. automodule:: pacman.lib.mpfit
     :members:
     :undoc-members:
@@ -140,21 +140,21 @@ lib.mpfit
 
 
 lib.geometry102
[…]
