### 2022-10-07 #4604 @xuanxu
@astrobel / @benmontet would you be willing to review this paper? It is a pipeline to analyze observations of transiting exoplanets written in Python that I think can fit with your expertise.
