+++ src/pacman/lib/mcmc.py
@@ -85,6 +85,12 @@ def mcmc_fit(data, model, params, file_name, meta, fit_par):
 
     plots.rmsplot(model, data, meta, fitter='mcmc')
 
+    if meta.s30_fit_white:
+        outfile = open(meta.workdir + meta.fitdir + '/white_systematics_mcmc.txt', "w")
+        for i in range(len(fit.all_sys)): print(fit.all_sys[i], file=outfile)
+        print('Saved white_systematics.txt file for mcmc run')
+        outfile.close()
+
     return medians, errors_lower, errors_upper, fit
 
 
