### 2022-10-19 992f358
added multiprocessoring for sampling, added some docstrings, fixed some flake8
