+++ src/pacman/s30_run.py
@@ -1,5 +1,4 @@
 import sys
-import matplotlib.pyplot as plt
 import numpy as np
 import glob
 import os
