+++ docs/source/stage00.rst
@@ -3,27 +3,46 @@
 Stage 00
 ============
 
+.. raw:: html
+
+    <style> .blue {color:blue} </style>
+
+.. topic:: TL;DR
+
+    - Decide on a run directory
+    - Copy pacman_script.py, fit_par.txt and obs_par.pcf to the run directory
+    - Enter path to run directory and data directory into the pcf
+    - Navigate to the run directory and execute the pacman_script.py file using the --s00 flag
+    - Continue with s01
+
+
 1) **Set up pcf**
 
-	First, set up the location of your data directory (=datadir) and run directory (=rundir).
+    .. role:: blue
+    :blue:`First, set up the location of your data directory (=datadir) and run directory (=rundir).`
 
-	The rundir should contain the following three files:
 
-	- **run_pacman.py**
+    For me, these directories are located at the following places:
 
-	- **fit_par.txt**
+    - datadir = /home/zieba/Desktop/Data/GJ1214_Hubble13021
+    - rundir = /home/zieba/Desktop/Projects/Observations/Hubble/GJ1214_13021
 
-	- **obs_par.pcf**
+    The rundir should contain the following three files:
 
-	They can be either found in your package directory, on `GitHub <https://github.com/sebastian-zieba/PACMAN/tree/master/src/pacman/data/run_files>`_
+    - **pacman_script.py**
+
+    - **fit_par.txt**
+
+    - **obs_par.pcf**
+
+	These files can be either found in your package directory, on `GitHub <https://github.com/sebastian-zieba/PACMAN/tree/master/src/pacman/data/run_files>`_
 	or can be downloaded `here <https://downgit.github.io/#/home?url=https://github.com/sebastian-zieba/PACMAN/tree/master/src/pacman/data/run_files>`_.
 
-	If your datadir contains a lot of visits, but you only want to analyze a subset you can use ``which_visits``.
-
 	As mentioned in the Introduction, in this example going to analyze just two
-	visits taken in the middle of the GO 13021 program (Dates (YYYY-MM-DD): 2013-03-13,  2013-03-15).
+	visits taken in the middle of the GO 13021 program (Dates (YYYY-MM-DD): 2013-03-13 and 2013-03-15).
 
-	If the user downloaded all 15 visits in GO 13021, he or she can choose: ``which_visits = [5,6]`` in the pcf.
+    If your datadir contains a lot of visits, but you only want to analyze a subset you can use the ``which_visits`` parameter in the pcf.
+    If the user downloaded all 15 visits in GO 13021, he or she can choose: ``which_visits = [5,6]`` in the pcf.
 
 
 2) **Run PACMAN**
@@ -38,6 +57,7 @@ Stage 00
 		python pacman_script.py --s00 --eventlabel='GJ1214_Hubble13021'
 
 	Here, --s00 means we are going to run Stage 00 and --eventlabel will be used in the naming of files and directories.
+    The  --eventlabel argument is just needed when running --s00.
 
 	When running s00, the first step creates a new subdirectory in rundir for the analysis which we will call the work directory (=workdir).
 
@@ -46,7 +66,7 @@ Stage 00
 
 	The fit_par.txt and obs_par.pcf files which are in the run directory will be copied there.
 
-	You have the following data structure now:
+	You should have a data structure similar to the following now:
 
 ::
 
