+++ docs/source/stage01.rst
@@ -3,6 +3,11 @@
 Stage 01
 ============
 
+.. topic:: TL;DR
+
+    - Navigate to the run directory and execute the pacman_script.py file using the --s01 flag
+    - Continue with s02
+
 Next we download the locations of HST. This will be later used for the barycentric correction.
 
     .. warning:: This step needs an internet connection!
@@ -10,6 +15,16 @@ Next we download the locations of HST. This will be later used for the barycentr
     .. note:: | At the beginning of every stage we read in again the pcf file located in the work directory.
               | This ensures that any user-made changes to the pcf file will be considered when running a new stage.
               | This means that the pcf file in the run directory is ONLY used in Stage 00. The same is true for the fit_par.txt file. So, after running Stage 00, PACMAN does not care anymore about the changes made to the pcf file and the fit_par file in the run directory.
+
+Navigate to your rundir (where pacman_script.py is located) in your terminal and type:
+
+.. code-block:: console
+
+    python pacman_script.py --s01
+
+The script (pacman_script.py) will assume that you want to continue the analysis started in the newly created work directory.
+If you want to use a different work directory instead, you can use the --workdir='SOME_PATH' flag when running the script.
+
 
 After running Stage 01 you should get an output like this:
 
@@ -28,7 +43,7 @@ Each file contains the X, Y and Z position of HST relative to the solar system b
 
 For example, the first of the two horizon files should look like this (due to its length, we just display the first 100 lines):
 
-.. include:: media/s01/horizons_results_v1_short.txt
+.. include:: media/s01/horizons_results_v0_short.txt
    :literal:
 
 The next Stage uses the information in these files to convert from MJD to BJD.
