+++ docs/source/media/s30/fit_par_white.txt
@@ -1,24 +1,0 @@
-#parameter    fixed   tied   value        lo_lim   lo_val   hi_lim   hi_val   prior   p1        p2       step_size  
-per           True    -1     1.58040433   False    1.58     False    1.581    X       2.25314   2e-05    1e-07      
-t0            False   -1     0.16         False    0.1      False    0.2      U       0.1       0.2      0.001      
-t_secondary   True    -1     0.0          False    0.0      False    0.0      X       0.0       0.0      0.0        
-w             True    -1     90.0         False    0.0      False    0.0      X       0.0       0.0      0.0        
-a             True    -1     14.85        False    3.0      False    10.0     U       3.0       10.0     0.02       
-inc           True    -1     88.7         False    75.0     False    89.99    U       75.0      89.99    0.02       
-rp            False   -1     0.15         True     0.03     True     0.4      U       0.01      0.4      0.001      
-fp            True    -1     0.0          False    0.0      False    0.002    X       0.0       1.0      0.0        
-u1            False   -1     0.27         True     0.0      True     0.99     U       0.0       1.0      0.01       
-u2            True    -1     0.0          False    0.0      False    0.0      X       0.0       0.0      0.0        
-ecc           True    -1     0.0          False    0.0      False    0.0      X       0.0       0.0      0.0        
-c             False   0      8.5          True     8.0      True     9.0      U       8.0       9.0      0.001      
-c             False   1      8.5          True     8.0      True     9.0      U       8.0       9.0      0.001      
-v             False   0      -1e-06       True     -0.001   True     -1e-09   U       -0.0001   0.0001   1e-08      
-v             False   1      -1e-06       True     -0.001   True     -1e-09   U       -0.0001   0.0001   1e-08      
-v2            True    -1     0.0          False    0.0      False    1.0      U       1e-10     1e-07    1e-11      
-r1            False   0      0.1          True     0.0      True     1.0      U       0.0001    20.0     0.01       
-r1            False   1      0.1          True     0.0      True     1.0      U       0.0001    20.0     0.01       
-r2            False   0      6.0          True     5.5      True     7.5      U       0.0001    20.0     0.1        
-r2            False   1      6.0          True     5.5      True     7.5      U       0.0001    20.0     0.1        
-r3            True    -1     0.0          False    -1.0     False    1.0      U       -10.0     10.0     0.001      
-scale         False   0      0.001577     True     -0.01    True     0.01     U       -0.1      0.1      0.002      
-scale         False   1      0.001577     True     -0.01    True     0.01     U       -0.1      0.1      0.002      

[…]
