+++ docs/source/pcf.rst
@@ -575,6 +575,11 @@ run_nested
 Runs nested sampling using the dynesty package.
 
 
+ncpu
+''''''''''''''''''''''''''''''''''''''''''''
+Number of cores used by emcee or dynesty. Use ncpu = 1, if you don't want to use parallelization.
+
+
 run_nsteps/run_nwalkers/run_nburn
 ''''''''''''''''''''''''''''''''''''''''''''
 Parameters for emcee.
