+++ src/pacman/__init__.py
@@ -1,8 +1,9 @@
-import pacman.s00_table
-import pacman.s01_horizons
-import pacman.s02_barycorr
-import pacman.s03_refspectra
-import pacman.s10_direct_images
-import pacman.s20_extract
-import pacman.s21_bin_spectroscopic_lc
-import pacman.s30_run
+from . import lib
+from . import s00_table
+from . import s01_horizons
+from . import s02_barycorr
+from . import s03_refspectra
+from . import s10_direct_images
+from . import s20_extract
+from . import s21_bin_spectroscopic_lc
+from . import s30_run
