### 2022-10-20 sebastian-zieba/PACMAN#6 @sebastian-zieba
Hi,
the problem actually just lies in the **fit_par.txt** file. The template I have uploaded to GitHub is based on an old version of the documentation where I showed the user how to reduce and analyze three visits at once.

In order to save a bit on computational time and disk space, I rewrote the docs a few weeks ago to show the reduction and analysis of just two visits instead. I forgot to update the **fit_par.txt** file though.

I just uploaded a new version of the **fit_par.txt** file to GitHub: https://github.com/sebastian-zieba/PACMAN/blob/master/src/pacman/data/run_files/fit_par.txt
If you replace your **fit_par.txt** with this new one, it should all work out.

I also tweaked a few parameters in the **pcf** so that the user can run the Quickstart a bit quicker. I just decided to reduce the number of iterations performed by _emcee_ when running this tutorial. If you want to save some time, you can adopt these changes too:

run_nsteps                   4000
run_nwalkers	             30
run_nburn                    2000

Hope this helps!
