+++ docs/source/conf.py
@@ -18,7 +18,7 @@ sys.path.insert(0, os.path.abspath('.'))
 sys.path.insert(0, os.path.abspath('../'))
 sys.path.insert(0, os.path.abspath('../../'))
 sys.path.insert(0, os.path.abspath('../../src/'))
-sys.path.insert(0, os.path.abspath('../../src/pacman/'))
+
 
 # -- Project information -----------------------------------------------------
 
