### 2022-10-20 d5c5fda
add missing dependency to docs
