### 2022-10-20 3b00d44
add missing dependency to docs
