+++ docs/requirements.txt
@@ -23,6 +23,6 @@ sphinx-gallery
 nbconvert
 ipython_genutils
 celerite
-
+exotic-ld
 
 
