### 2022-10-20 #4838 @astrobel
Hi @sebastian-zieba, thanks for the great software package! PACMAN is intuitive for the new user and covers a lot of ground really efficiently, and on these grounds I'm happy to recommend it for publication in JOSS. I have several notes below which I hope can help improve the documentation and the example case in particular.

### General notes

* I'm confused by the `pip install .[test]` line and instructions in the installation section of the docs. Nevertheless, by installing pytest in my conda environment I'm able to run the tests, which pass with 36 warnings. I don't think these warnings need to be addressed for acceptance to JOSS, but I'd be happy to send them your way if you're interested.

* If running Stage 01 in a directory that contains a Jupyter Notebook file & associated directory, `.ipynb_checkpoints`, Stage 01 detects that as the run directory and exits with the following output:

```
workdir:  ./.ipynb_checkpoints/
eventlabel:
Traceback (most recent call last):
  File "/home/isy/aur/PACMAN/src/pacman/pacman_script.py", line 135, in <module>
    main()
  File "/home/isy/aur/PACMAN/src/pacman/pacman_script.py", line 106, in main
    update_meta(eventlabel, workdir)
  File "/home/isy/aur/PACMAN/src/pacman/lib/update_meta.py", line 14, in update_meta
    meta = me.loadevent(workdir + '/WFC3_' + eventlabel + "_Meta_Save")
  File "/home/isy/aur/PACMAN/src/pacman/lib/manageevent.py", line 143, in loadevent
    handle = open(filename + '.dat', 'rb')
FileNotFoundError: [Errno 2] No such file or directory: './.ipynb_checkpoints//WFC3__Meta_Save.dat'
```

* I was able to fix this by moving to a clean directory. As it was an easy fix, I haven't opened an issue, but let me know if you want one opened for it!

* In the Quickstart tutorial, the following parameters are missing from Stage 20 in the example PCF:

```
background_box               False
bg_rmin                      100
bg_rmax                      400
bg_cmin                      40
bg_cmax                      100

correct_wave_shift_refspec   True
```

* Also, the PCF default for Stage 20 doesn't save any of the figures from the Quickstart example, I think it's worth changing that or at least highlighting the relevant PCF lines in the docs.

* And in Stage 30, the example PCF is missing the following:

```
remove_which_orb            [0]

ld_model                     1
```

* Please include community guidelines for third parties wishing to contribute to the software, preferably in the readme on the PACMAN landing page!

### Documentation

* The links to the individual stages on the "Stages" page are broken.

* Under "Repository Structure", it might be worth noting that the user should move the run files to their working directory. To keep things clear and separate, I suggest not including lines of example code in the 'Repository Structure" section, and keeping all code examples to the Quickstart section. Where necessary, you can point a link to a relevant section of the tutorial from the structural sections of the documentation.

* In the quickstart tutorial, setting up the directory structure should be included before the download instructions, as that's foundational to how PACMAN works. I strongly suggest incorporating a step that downloads/moves the ima files directly into the user's "data" directory.

### Paper

The paper is great, and very thorough! I only have a couple of notes:

* I think you could flesh out the state of the field a bit more. Are there other commonly-used pipelines that perform parts of what PACMAN can accomplish all-in-one? This could be included in the statement of need section where you mention that other pipelines have produced discrepant results.

* Please clarify that it's GJ1214 shown in the example Figures 1 and 2. Are there other papers on this object or are these new results? This will also help clarify the reproducability of these results.
