### 2022-10-28 f402451
JOSS review: update docs and README.md
