+++ docs/source/stage03.rst
@@ -3,7 +3,7 @@
 Stage 03
 ============
 
-.. topic:: TL;DR
+.. topic:: Summary
 
     - Navigate to the run directory and execute the pacman_script.py file using the --s03 flag
     - Continue with s10
