### 2022-10-28 13bc286
fixed DepricationWarning in test_all.py
