### 2022-10-28 #4838 @sebastian-zieba
Hi @astrobel,

thanks again for your comments!
I’m still working on the JOSS paper, but I’ve already incorporated all of the other code or documentation-related comments. Here are my replies to your individual comments:

> ### General notes
> * I'm confused by the `pip install .[test]` line and instructions in the installation section of the docs. Nevertheless, by installing pytest in my conda environment I'm able to run the tests, which pass with 36 warnings. I don't think these warnings need to be addressed for acceptance to JOSS, but I'd be happy to send them your way if you're interested.

The `[test]` argument will additionally install the testing requirements (e.g., pytest). If you just do `pip install -e .` it will install the necessary PACMAN requirements (like astropy, numpy, …) but not pytest.
I added more explanation on the `[test]` (and the `[docs]`) arguments in the installing instructions.

Here’s the text i added to https://pacmandocs.readthedocs.io/en/latest/installation.html:

.. note:: ``pip install -e .`` will only install the necessary dependencies. If you also want to install the dependencies to run the tests (using pytest) or work on the docs, you have to use the ``[test]`` or ``[docs]`` arguments respectively after ``pip install -e .``. See `Test Your Installation <https://pacmandocs.readthedocs.io/en/latest/installation.html#test-your-installation>`_ for an example.

I also added the following further down:

.. code-block:: console

   pip install .[test]

The ``[test]`` argument will also install the necessary dependencies to run pytest.

Regarding the tests: I am also getting some warnings when running the tests. I will work on these in the upcoming releases. Thanks for pointing this out.

> * If running Stage 01 in a directory that contains a Jupyter Notebook file & associated directory, `.ipynb_checkpoints`, Stage 01 detects that as the run directory and exits with the following output:
>
> ```
> workdir:  ./.ipynb_checkpoints/
> eventlabel:
> Traceback (most recent call last):
>   File "/home/isy/aur/PACMAN/src/pacman/pacman_script.py", line 135, in <module>
>     main()
>   File "/home/isy/aur/PACMAN/src/pacman/pacman_script.py", line 106, in main
>     update_meta(eventlabel, workdir)
>   File "/home/isy/aur/PACMAN/src/pacman/lib/update_meta.py", line 14, in update_meta
>     meta = me.loadevent(workdir + '/WFC3_' + eventlabel + "_Meta_Save")
>   File "/home/isy/aur/PACMAN/src/pacman/lib/manageevent.py", line 143, in loadevent
>     handle = open(filename + '.dat', 'rb')
> FileNotFoundError: [Errno 2] No such file or directory: './.ipynb_checkpoints//WFC3__Meta_Save.dat'
> ```
>
> * I was able to fix this by moving to a clean directory. As it was an easy fix, I haven't opened an issue, but let me know if you want one opened for it!

That’s a great point, thanks!
I changed the `pacman_script.py` file. Stage 01 (and the later stages) will therefore ignore directories in the run directory that don't have names starting with "run_".

> * In the Quickstart tutorial, the following parameters are missing from Stage 20 in the example PCF:
>
> ```
> background_box               False
> bg_rmin                      100
> bg_rmax                      400
> bg_cmin                      40
> bg_cmax                      100
>
> correct_wave_shift_refspec   True
> ```
>

Fixed

> * Also, the PCF default for Stage 20 doesn't save any of the figures from the Quickstart example, I think it's worth changing that or at least highlighting the relevant PCF lines in the docs.

Fixed

> * And in Stage 30, the example PCF is missing the following:
>
> ```
> remove_which_orb            [0]
>
> ld_model                     1
> ```
>

Fixed

> * Please include community guidelines for third parties wishing to contribute to the software, preferably in the readme on the PACMAN landing page!

I added some information on how to contribute to the README on GitHub.

>
> ### Documentation
> * The links to the individual stages on the "Stages" page are broken.

Thanks!! I actually removed the links there. They were links to the source code of the individual stages. I think that makes it more confusing and the user can instead just check out the code on GitHub (if they want).

> * Under "Repository Structure", it might be worth noting that the user should move the run files to their working directory. To keep things clear and separate, I suggest not including lines of example code in the 'Repository Structure" section, and keeping all code examples to the Quickstart section. Where necessary, you can point a link to a relevant section of the tutorial from the structural sections of the documentation.

Thanks! I agree. I removed the code in Repository Structure and noted clearly now that the user has to copy the contents of run_files into the work directory.

> * In the quickstart tutorial, setting up the directory structure should be included before the download instructions, as that's foundational to how PACMAN works. I strongly suggest incorporating a step that downloads/moves the ima files directly into the user's "data" directory.

Yes! I worked on making the whole Before Running and Stage 00 Section clearer.
I have the following order for the Quickstart now:
Introduction to Quickstart -> Setting up directories (data and run directory) -> download data and move to data directory -> start with stage 00.

I'll let you know when I also pushed the changes related to the paper.
