+++ joss/paper.bib
@@ -504,6 +504,17 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@MISC{Bean2012,
+       author = {{Bean}, Jacob},
+        title = "{Revealing the Diversity of Super-Earth Atmospheres}",
+ howpublished = {HST Proposal ID 13021. Cycle 20},
+         year = 2012,
+        month = oct,
+        pages = {13021},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2012hst..prop13021B},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
 
 
 
