+++ joss/paper.md
@@ -107,11 +107,11 @@ scans, $r_{\rm{1}}$ and $r_{\rm{2}}$ are parameters to account for the exponenti
 \autoref{fig:figure1} (right panel) shows the resulting 1D spectrum and a user-defined binning.
 
 \autoref{fig:figure1} and \autoref{fig:figure2} show some figures created by PACMAN during a run using three HST visits of GJ 1214 b collected in [GO 13201](https://archive.stsci.edu/proposal_search.php?id=13021&mission=hst) [@Bean2012].
+An analysis of all visits was published in @Kreidberg2014.
 
+![_Left panel_: a typical single exposure showing the raw 2D spectrum. _Right panel_: 1D spectrum after the use of optimal extraction including vertical dashed lines showing the user-set binning to generate spectroscopic light curves.\label{fig:figure1}](figures/fig1.pdf "title-2"){ width=99.9% }
 
-![_Left panel_: a typical single exposure showing the raw 2D spectrum. _Right panel_: 1D spectrum after the use of optimal extraction including vertical dashed lines showing the user-set binning to generate spectroscopic light curves.\label{fig:figure1}](figures/fig1.pdf "title-2"){ width=99% }
-
-![_panel A_: raw white light curves for each visit. One can clearly see the constant offset between two adjacent exposured due to the spatial scanning mode. _panel B_: white light curve with the best astrophysical model fit using \autoref{eq:equation1}. _panel C_: the transmission spectrum after fitting 11 spectroscopic light curves.\label{fig:figure2}](figures/tmpfig2.pdf "title-2"){ width=99% }
+![_panel A_: raw white light curves for each of the three visits. One can clearly see the constant offset between two adjacent exposures due to the spatial scanning mode. _panel B_: white light curve with the best astrophysical model fit using \autoref{eq:equation1}. _panel C_: the transmission spectrum after fitting 11 spectroscopic light curves reveiling the flat spectrum of GJ 1214 b as published in @Kreidberg2014.\label{fig:figure2}](figures/fig2.pdf "title-2"){ width=99.9% }
 
 
 # Dependencies
