+++ joss/paper.bib
@@ -515,6 +515,112 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{Eastman2019,
+       author = {{Eastman}, Jason D. and {Rodriguez}, Joseph E. and {Agol}, Eric and {Stassun}, Keivan G. and {Beatty}, Thomas G. and {Vanderburg}, Andrew and {Gaudi}, B. Scott and {Collins}, Karen A. and {Luger}, Rodrigo},
+        title = "{EXOFASTv2: A public, generalized, publication-quality exoplanet modeling code}",
+      journal = {arXiv e-prints},
+     keywords = {Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics},
+         year = 2019,
+        month = jul,
+          eid = {arXiv:1907.09480},
+        pages = {arXiv:1907.09480},
+archivePrefix = {arXiv},
+       eprint = {1907.09480},
+ primaryClass = {astro-ph.EP},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2019arXiv190709480E},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Luger2019,
+       author = {{Luger}, Rodrigo and {Agol}, Eric and {Foreman-Mackey}, Daniel and {Fleming}, David P. and {Lustig-Yaeger}, Jacob and {Deitrick}, Russell},
+        title = "{starry: Analytic Occultation Light Curves}",
+      journal = {\aj},
+     keywords = {eclipses, methods: analytical, occultations, techniques: photometric, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Solar and Stellar Astrophysics},
+         year = 2019,
+        month = feb,
+       volume = {157},
+       number = {2},
+          eid = {64},
+        pages = {64},
+          doi = {10.3847/1538-3881/aae8e5},
+archivePrefix = {arXiv},
+       eprint = {1810.06559},
+ primaryClass = {astro-ph.IM},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2019AJ....157...64L},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@MISC{Gunther2019,
+       author = {{G{\"u}nther}, Maximilian N. and {Daylan}, Tansu},
+        title = "{allesfitter: Flexible star and exoplanet inference from photometry and radial velocity}",
+     keywords = {Software},
+ howpublished = {Astrophysics Source Code Library, record ascl:1903.003},
+         year = 2019,
+        month = mar,
+          eid = {ascl:1903.003},
+        pages = {ascl:1903.003},
+archivePrefix = {ascl},
+       eprint = {1903.003},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2019ascl.soft03003G},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Gunther2021,
+       author = {{G{\"u}nther}, Maximilian N. and {Daylan}, Tansu},
+        title = "{Allesfitter: Flexible Star and Exoplanet Inference from Photometry and Radial Velocity}",
+      journal = {\apjs},
[…]
