+++ joss/paper.md
@@ -95,6 +95,7 @@ Here some examples for the currently implemented models for the instrument syste
 
 A typical model to fit an exoplanet transit in HST data is the following (used for example in @Kreidberg2014a): 
 
+
     \begin{equation}
     \label{eq:equation1}
     F(t) = T(t) \, (c\,S(t) + k\,t_{\rm{v}}) \, (1 - \exp(-r_1\,t_{\rm{orb}} - r_2 )),
