+++ joss/paper.md
@@ -92,6 +92,8 @@ Here some examples for the currently implemented models for the instrument syste
     - transit and secondary eclipse curves as implemented in `batman`
     - sinusoids for phase curve fits
     - a constant offset which accounts for the upstream-downstream effect [@McCullough2012] caused by forward and reverse scanning
+    
+ A typical model to fit an exoplanet transit in HST data is the following (used for example in @Kreidberg2014a): 
 
 
     \begin{equation}
