+++ joss/paper.md
@@ -54,7 +54,7 @@ This will enable easy comparison between different pipelines, and also lower the
 
 What sets `PACMAN` apart from other tools provided by the community, is that it was specifically designed to reduce and fit HST data.
 For example, there are several open-source tools which can fit time series observations 
-of stars to model events like transiting exoplanets (e.g., EXOFASTv2 [@Eastman2019], juliet [@Espinoza2019], allesfitter [@Gunther2019; @Gunther2021], exoplanet [@Foreman-Mackey2021a; @Foreman-Mackey2021b], starry [@Luger2019]). 
+of stars to model events like transiting exoplanets (e.g., `EXOFASTv2` [@Eastman2019], `juliet` [@Espinoza2019], `allesfitter` [@Gunther2019; @Gunther2021], `exoplanet` [@Foreman-Mackey2021a; @Foreman-Mackey2021b], `starry` [@Luger2019]). 
 `PACMAN`'s source code, however, includes fitting models which can model systematics which are characteristic to HST data 
 (like the orbit-long exponential ramps due to charge trapping or the upstream-downstream effect)
 which removes the need for the user to write this functions themselves.
@@ -81,7 +81,7 @@ We also mask bad pixels that have been flagged by `calwf3` with data quality DQ 
 The _ima_ files taken in this observation mode consist of a number of nondestructive reads, also known as up-the-ramp samples, each of which we treat as an independent subexposure.
 \autoref{fig:figure1} (left panel) shows an example for the last subexposure when using the spatial scanning together with the expected position of the trace based on the direct image.
 
-- **Fitting models**: `PACMAN` contains several functions to fit models which are commently used with HST data. 
+- **Fitting models**: `PACMAN` contains several functions to fit models which are commonly used with HST data. 
 The user can fit models like in \autoref{eq:equation1} to the white light curve or to spectroscopic light curves. 
 An example for a raw spectroscopic light curve and fitting \autoref{eq:equation1} to it, can be found in \autoref{fig:figure2}.
 Here some examples for the currently implemented models for the instrument systematics and the astrophysical signal.
@@ -92,8 +92,10 @@ Here some examples for the currently implemented models for the instrument syste
     - transit and secondary eclipse curves as implemented in `batman`
     - sinusoids for phase curve fits
     - a constant offset which accounts for the upstream-downstream effect [@McCullough2012] caused by forward and reverse scanning
-    
-  A typical model to fit an exoplanet transit in HST data is the following (used for example in @Kreidberg2014a): 
+
+
+
+A typical model to fit an exoplanet transit in HST data is the following (used for example in @Kreidberg2014a):
 
 
     \begin{equation}
@@ -105,10 +107,10 @@ Here some examples for the currently implemented models for the instrument syste
 with _T(t)_ being the transit model, _c_ (_k_) a constant (slope), _S(t)_ a scale factor equal to 1 for exposures with spatial scanning in the forward direction and _s_ for reverse
 scans, $r_{\rm{1}}$ and $r_{\rm{2}}$ are parameters to account for the exponential ramps. $t_{\rm{v}}$ and $t_{\rm{orb}}$ are the times from the first exposure in the visit and in the orbit, respectively.
 
-- **parameter estimation**: The user has different options to estimate best fitting parameters and their uncertainies:
+- **parameter estimation**: The user has different options to estimate best fitting parameters and their uncertainties:
   - least square: `scipy.optimize`
-  - MCMC: `emcee`
-  - nested sampling: `dynesty`
+  - MCMC: `emcee` [@Foreman-Mackey2013]
+  - nested sampling: `dynesty` [@Speagle2020]
 
 - **multi-visit observations**
[…]
