### 2022-11-02 #4838 @sebastian-zieba
Hi @astrobel ,

I just pushed the last changes related to the paper. I have therefore now addressed all your comments and incorporated your suggested changes to the code, documentation and paper.

Here are my changes to the paper:

> * Please clarify that it's GJ1214 shown in the example Figures 1 and 2. Are there other papers on this object or are these new results? This will also help clarify the reproducability of these results.

I updated this part and clearly state now that the plots I am showing in the paper only use three Hubble-visits of the planet GJ 1214 b. I reference the paper Kreidberg+2014 which analyzed all 15 visits. The plots I am showing are therefore not new results but a reanalysis of public data. The results using PACMAN are also consistent with the published results.

(I also updated the figures 1 & 2 to make them a bit prettier.)

> * I think you could flesh out the state of the field a bit more. Are there other commonly-used pipelines that perform parts of what PACMAN can accomplish all-in-one? This could be included in the statement of need section where you mention that other pipelines have produced discrepant results.

I also updated the statement of need listing some examples by how PACMAN sets itself apart from other codes.

Best, Sebastian
