### 2022-11-16 #4838 @xuanxu
@grburgess Friendly reminder.
Please create your checklist with the following command: `@editorialbot generate my checklist`
