### 2022-12-13 #4838 @xuanxu
Hey @aureliocarnero: thanks for agreeing to review!

As a first step, please create your checklist with the following command: `@editorialbot generate my checklist`
