### 2022-12-13 #4838 @aureliocarnero
El mar, 13 dic 2022, 15:22, Juanjo Bazán ***@***.***>
escribió:

> Hey @aureliocarnero <https://github.com/aureliocarnero>: thanks for
> agreeing to review!
>
> As a first step, please create your checklist with the following command: @editorialbot
> generate my checklist
>
