+++ docs/source/media/s30/spectroscopic/fit_results.txt
@@ -1,0 +1,15 @@
+===== ======== ======= ====== ======= ======== ======== ================ =================== =============== =================== ======= =========== ===
+ wave rms_pred     rms   chi2 chi2red      bic  ln_like chi2_notrescaled chi2red_notrescaled bic_notrescaled ln_like_notrescaled npoints nfree_param dof
+===== ======== ======= ====== ======= ======== ======== ================ =================== =============== =================== ======= =========== ===
+1.158  247.257 284.441  93.05   0.979 1711.148  -825.14          143.033               1.506        1714.698            -826.915     108          13  95
+1.204  229.721 250.839  89.47   0.942 1714.542 -826.837          128.622               1.354        1714.493            -826.813     108          13  95
+1.250  220.843 212.878 93.465   0.984 1695.436 -817.284          100.451               1.057        1694.637            -816.885     108          13  95
+1.296  212.792 221.103 87.954   0.926 1719.518 -829.325          116.482               1.226        1717.707             -828.42     108          13  95
+1.342  213.148 222.264 89.804   0.945 1719.704 -829.418          117.307               1.235        1718.353            -828.743     108          13  95
+1.389  219.237 229.906 90.108   0.949 1716.892 -828.012           118.46               1.247        1715.699            -827.415     108          13  95
+1.435  220.256 249.262  93.53   0.985 1730.668   -834.9          138.251               1.455        1733.184            -836.158     108          13  95
+1.481  222.435 236.198  93.63   0.986 1714.989  -827.06           121.62                1.28        1714.731            -826.932     108          13  95
+1.527  221.682  215.73 93.335   0.982 1697.484 -818.308          102.429               1.078        1696.537            -817.835     108          13  95
+1.573  222.021 215.512 89.944   0.947 1696.632 -817.882          101.657                1.07        1695.124            -817.128     108          13  95
+1.619  226.484  243.15 90.961   0.957  1714.84 -826.986          124.507               1.311        1714.482            -826.807     108          13  95
+===== ======== ======= ====== ======= ======== ======== ================ =================== =============== =================== ======= =========== ===
