+++ docs/source/outputs.rst
@@ -1,0 +1,333 @@
+.. _Outputs:
+
+Outputs
+=======
+
+Here we summarize all the files created by PACMAN when doing a complete run from s00 to s30.
+
+
+Stage 00
+''''''''
+
+directories
+-----------
+- workdir (e.g.: run_2022-12-16_10-46-31_dec2022) 
+    the main directory of your analysis
+
+- workdir/figs
+    all figures will be saved into here
+
+- workdir/figs/s00_obs_dates
+    
+
+plots
+------------
+- workdir/figs/s00_obs_dates/obs_dates.png
+- workdir/figs/s00_obs_dates/obs_dates_all.png
+
+other files
+------------
+
+- workdir/WFC3_dec2022_Meta_Save.dat
+    the meta file: this file will collect a lot of useful information when finishing a stage and then accessed by the next stage.
+
+- workdir/filelist.txt
+    a list containing all *ima.fits files in the datadir and some information from their headers.
+
+- workdir/fit_par.txt
+    this fit_par.txt will be used for the fitting in s30
+
+- workdir/obs_par.pcf
+    this is the main obs_par.pcf file for all future stages
+
+Stage 01
+''''''''
+
+directories
+-----------
+- ancil
+- ancil/horizons
+
+other files
+------------
+
+- ancil/horizons/horizons_results_v0.txt
+
+
+Stage 02
+''''''''
+
+directories
+-----------
+- workdir/figs/s02_barycorr
+
+plots
+------------
+- workdir/figs/s02_barycorr/bjdcorr_horizons_results_v0.png
+
+
+Stage 03
+''''''''
+
+directories
+-----------
+
+- workdir/ancil/refspec
+
+- workdir/ancil/stellar_models
+
+- workdir/figs/s03_refspec
+
+- workdir/figs/s03_smooth
+
+plots
+------------
+
+- workdir/figs/s03_refspec/refspec.png
+
+- workdir/figs/s03_smooth/smooth.png
+
+other files
+------------
+
+- workdir/ancil/stellar_models/k93models/kp03_3500.fits
+
+
+
+Stage 10
+''''''''
+
+directories
+-----------
+
+- workdir/figs/s10_images
+
+plots
+------------
+
+- workdir/figs/di_0.png
+- workdir/figs/quick_di0.png
+
+other files
+------------
+
+- workdir/xrefyref.txt
+
+
+
+
+Stage 21
+''''''''
+
+directories
+-----------
+
+- workdir/extracted_lc/2022-12-16_11-22-12
+- workdir/figs/s20_badmask
+- workdir/figs/s20_bkg_evo
+- workdir/figs/s20_bkg_hist
+- workdir/figs/s20_drift
+- workdir/figs/s20_optextr
+- workdir/figs/s20_refspec_fit
+- workdir/figs/s20_sp1d
+- workdir/figs/s20_sp1d_diff
+- workdir/figs/s20_sp2d
+- workdir/figs/s20_trace
+- workdir/figs/s20_utr
+- workdir/figs/s20_utr_aper_evo
+
+plots
+------------
+
+- workdir/figs/s20_badmask/badmask_0
+- workdir/figs/s20_bkg_evo/bkg_evo.png
+- workdir/figs/s20_bkg_hist/bkg_hist0-0.png
[…]
