+++ docs/source/stage30.rst
@@ -1,4 +1,4 @@
-.. _stage31:
+.. _stage30:
 
 Stage 30
 ============

[…]
