### 2022-12-16 7f4e6ff
fixed typos in joss paper
