+++ joss/paper.md
@@ -30,14 +30,14 @@ bibliography: paper.bib
 The Hubble Space Telescope (HST) has become the preeminent workhorse facility for the characterization of extrasolar planets.
 Launched in 1990 and never designed for the observations of exoplanets, the STIS spectrograph on HST was used in 2002 to detect the first atmosphere ever discovered on a planet outside of our solar system [@Charbonneau2002].
 
-HST has currently the two most powerful operating tools in space to characterize exoplanets over a broad spectral range:
+HST has currently two of the most powerful operating tools in space to characterize exoplanets over a broad spectral range:
 The Space Telescope Imaging Spectrograph (STIS; installed in 1997) in the UV and the Wide Field Camera 3 (WFC3; installed in 2009) in the Near Infrared (NIR).
 With the introduction of a spatial scan mode on WFC3 [@McCullough2012; @Deming2012] where the star moves perpendicular to the dispersion direction during an exposure, WFC3 observations have become very efficient due to the reduction of overhead time and the possibility of longer exposures without saturation.
 
 For exoplanet characterization, WFC3 is used for transit and secondary eclipse spectroscopy and phase curve observations.
 The instrument has two different grisms: G102 with a spectral range from 800 nm to up to 1150 nm and G141 encompassing 1075 nm to about 1700 nm.
 The spectral range of WFC3/G141 is primarily sensitive to molecular absorption from water at approximately 1.4 microns.
-This lead to the successful detection of water in the atmosphere of over a dozen of exoplanets [e.g., @Deming2013; @Huitson2013; @Fraine2014; @Kreidberg2014b; @Evans2016].
+This led to the successful detection of water in the atmosphere of over a dozen of exoplanets [e.g., @Deming2013; @Huitson2013; @Fraine2014; @Kreidberg2014b; @Evans2016].
 The bluer part of WFC3, the G102 grism, is also sensitive to water and most notably led to the first detection of a helium exosphere [@Spake2018].
 
 Here we present `PACMAN`, an end-to-end pipeline developed to reduce and analyze HST/WFC3 data.
@@ -57,7 +57,7 @@ For example, there are several open-source tools which can fit time series obser
 of stars to model events like transiting exoplanets (e.g., `EXOFASTv2` [@Eastman2019], `juliet` [@Espinoza2019], `allesfitter` [@Gunther2019; @Gunther2021], `exoplanet` [@Foreman-Mackey2021a; @Foreman-Mackey2021b], `starry` [@Luger2019]). 
 `PACMAN`'s source code, however, includes fitting models which can model systematics which are characteristic to HST data 
 (like the orbit-long exponential ramps due to charge trapping or the upstream-downstream effect)
-which removes the need for the user to write this functions themselves.
+which removes the need for the user to write these functions themselves.
 `PACMAN` will also retrieve information from the header of the fits files, automatically detect HST orbits and visits and use this information in the fitting models.
 
 # Outline of the pipeline steps
@@ -77,14 +77,14 @@ This first exposure in the visit is then used as the template for the following 
 - **Optimal extraction and outlier removal**: `PACMAN` uses an optimal extraction algorithm as presented in @Horne1986 which iteratively masks bad pixels in the image. 
 We also mask bad pixels that have been flagged by `calwf3` with data quality DQ = 4 or 512\footnote{for a list of DQ flags see \url{https://wfc3tools.readthedocs.io/en/latest/wfc3tools/calwf3.html}}.
 
-- **Scanning of the detector**: The majority of exoplanetary HST/WFC3 observations use the spatial scanning technique [@McCullough2012] which spreads the light perpendicular to the dispersion direction during the exposure enabled longer integration times before saturation.
[…]
