+++ joss/paper.md
@@ -60,6 +60,9 @@ of stars to model events like transiting exoplanets (e.g., `EXOFASTv2` [@Eastman
 which removes the need for the user to write these functions themselves.
 `PACMAN` will also retrieve information from the header of the fits files, automatically detect HST orbits and visits and use this information in the fitting models.
 
+As mentioned above, there are several open-source codes that fit light curves but `PACMAN` also does data reduction. 
+Another code which can reduce HST/WFC3 observations is `Iraclis` [@Tsiaras2016] (more in Section  [Similar tools](#similar-tools)).
+
 # Outline of the pipeline steps
 
 The pipeline starts with the _ima_ data products provided by the Space Telescope Science Institute which can be easily accessed from [MAST](https://mast.stsci.edu/search/hst).
