+++ joss/metadata.yaml
@@ -10,7 +10,7 @@ variables:
 
 metadata:
   logo_path: "figures/joss-logo.png"
-  citation_author: "Sebastian Zieba et al."
+  citation_author: "Zieba, & Kreidberg"
   issue: "xxx"
   page: "xxx"
   volume: "xxx"
