### 2022-12-20 f632bb6
fixed typos in JOSS paper
