### 2022-12-20 #4838 @sebastian-zieba
@xuanxu Thanks for your work!
I created a new tag on GitHub (v0.3).
The new Zenodo DOI is 10.5281/zenodo.7465579
I checked that the authors on the JOSS paper are the same as the ones listed on Zenodo. I also added the ORCIDs to the Zenodo release.
