### 2022-12-21 #4838 @sebastian-zieba
I did also notice a typo in the title of the paper. I corrected it in the new tagged version (v0.3) (the one I'm referring to in the earlier post).
(Instead of "...Camera 3 IF Grism data" it's now "...Camera 3 IR Grism data").
I just wanted to point it out, to make sure it's correct in the final version.
