### 2022-12-21 a7d7f6d
updating journal macros
