+++ joss/paper.bib
@@ -1,7 +1,7 @@
 @ARTICLE{Kreidberg2014a,
        author = {{Kreidberg}, Laura and {Bean}, Jacob L. and {D{\'e}sert}, Jean-Michel and {Benneke}, Bj{\"o}rn and {Deming}, Drake and {Stevenson}, Kevin B. and {Seager}, Sara and {Berta-Thompson}, Zachory and {Seifahrt}, Andreas and {Homeier}, Derek},
         title = "{Clouds in the atmosphere of the super-Earth exoplanet GJ1214b}",
-      journal = {\nat},
+      journal = {Nature},
      keywords = {Astrophysics - Earth and Planetary Astrophysics},
          year = 2014,
         month = jan,
@@ -19,7 +19,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Kreidberg2018,
        author = {{Kreidberg}, Laura and {Line}, Michael R. and {Parmentier}, Vivien and {Stevenson}, Kevin B. and {Louden}, Tom and {Bonnefoy}, Mick{\"a}el and {Faherty}, Jacqueline K. and {Henry}, Gregory W. and {Williamson}, Michael H. and {Stassun}, Keivan and {Beatty}, Thomas G. and {Bean}, Jacob L. and {Fortney}, Jonathan J. and {Showman}, Adam P. and {D{\'e}sert}, Jean-Michel and {Arcangeli}, Jacob},
         title = "{Global Climate and Atmospheric Composition of the Ultra-hot Jupiter WASP-103b from HST and Spitzer Phase Curve Observations}",
-      journal = {\aj},
+      journal = {Astronomical Journal},
      keywords = {planets and satellites: atmospheres, planets and satellites: gaseous planets, planets and satellites: individual: WASP-103b, Astrophysics - Earth and Planetary Astrophysics},
          year = 2018,
         month = jul,
@@ -38,7 +38,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Kreidberg2015,
        author = {{Kreidberg}, Laura},
         title = "{batman: BAsic Transit Model cAlculatioN in Python}",
-      journal = {\pasp},
+      journal = {Publications of the ASP},
      keywords = {Astrophysics - Earth and Planetary Astrophysics},
          year = 2015,
         month = nov,
@@ -68,7 +68,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Horne1986,
        author = {{Horne}, K.},
         title = "{An optimal extraction algorithm for CCD spectroscopy.}",
-      journal = {\pasp},
+      journal = {Publications of the ASP},
      keywords = {Astronomical Spectroscopy, Charge Coupled Devices, Data Reduction, Spectrum Analysis, Image Processing, Iteration, Optimization, Spatial Resolution, Stellar Spectra, Variance (Statistics), Instrumentation and Photography},
          year = 1986,
         month = jun,
@@ -82,7 +82,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Foreman-Mackey2013,
        author = {{Foreman-Mackey}, Daniel and {Hogg}, David W. and {Lang}, Dustin and {Goodman}, Jonathan},
         title = "{emcee: The MCMC Hammer}",
-      journal = {\pasp},
+      journal = {Publications of the ASP},
      keywords = {Astrophysics - Instrumentation and Methods for Astrophysics, Physics - Computational Physics, Statistics - Computation},
          year = 2013,
[…]
