### 2022-12-21 #4838 @dfm
I've made some small edits to the manuscript and bibliography: https://github.com/sebastian-zieba/PACMAN/pull/8
Take a look at that, and then I can proceed with publication.
