### 2022-12-21 #4838 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
